final optimizations
