+++ src/core/dom.js
@@ -56,12 +56,9 @@ class Dom {
   }
 
   css(styles = {}) {
-    for (const key in styles) {
-      if ({}.hasOwnProperty.call(styles, key)) {
-        console.log(key);
-        console.log(styles[key]);
-      }
-    }
+    Object.keys(styles).forEach((key) => {
+      this.$el.style[key] = styles[key];
+    });
   }
 }
